Fix error which causes exports of binary GeoDB to be corrupted
Offset was not updated while downloading binary exports of the location
database.

This is a critical error since it causes downloads of binary exports to be
corrupted and missing data. Binary exports created with versions of
location-visualizer before v1.2.1 should NOT be used for backups, since they
will likely contain invalid data.